First (non-working) version
This uses a PUT_W_SYNCFLAG for the Long payload (upto some limit).
With the syncflag in the notify ring the ordering of payload before
header is automagical.  HOWEVER, this doesn't work since the put
and sync aren't in the same memory registration :-(

Placing a second ring in the auxseg would fix the registation issue,
but only for LARGE mode, and would NOT provide ordering - just the
ability to do matching after reception of the header.