ibv-conduit: checkpoint/restart support
With this commit we can checkpoint and restart ibv-conduit.  Initially
only self-initiated collective checkpoints (and collective restarts)
are supported.

This has been tested by taking a checkpoint right after attach in each
of the gasnet tests.  There were no errors in taking the chechpoint,
and no errors running the test after the checkpoint had been taken.
This supports, but does not prove, that all "problematic" ibv
resources where correctly released and reacquired.  The testing of
restarts has not been quite as complete, but testgasnet has been
restarted from the beginning of each of its "doitN" sub-divisions.