ibv-conduit: checkpoint+continue proof-of-concept
With this commit we can checkpoint ibv-conduit and continue afterward.

This has been tested by taking a checkpoint right after attach in each
of the gasnet tests.  There were no errors in taking the chechpoint,
and no errors running the test after the cehckpoint had been taken.
This supports, but does not prove, that all "probelematic" ibv
resources where correctly released and reacquired.