ibv-conduit: avoid pthread resource leaks/races
With the pending work for checkpoint/restart it becomes important that
when terminating a progress thread (which we will restart) we should
avoid leaking resources.  So this commit ensures that for each thread
we "stop" we will also issue a pthread_join() or pthread_detach() call
(which one depends on a new "block" argument).

The blocking case (join) will also help prevent races that could have
occured with the previous relaxed approach to thread termination.