ssh-spawner: improved BLCR support
This commit adds (conduit-independent) pre- and post-checkpoint
callbacks in the ssh-spawner code, and invokes them in ibv-conduit.
This allows ssh-spawner to handle a couple minor items that were
(knowingly) ignored in the initial implementation.  Specifically,
prior to this commit there were two known issues with any restarted
process:

1) The control socket's close-on-exec flag was not set, and thus the
socket would still be open in any childen of a restarted process.  Now
the close-on-exec flag is set as it would have been in the absence of
checkpoint/restart.

2) The table of pids maintained by the master process would contain
the pids of the cr_restart processes rather than the GASNet processes.
This could have been a problem when sending SIGKILL in the case of
abnormal termination[1].  Now the gather_pids() step required in each
GASNet process will occur after the restart rather than before.

The new callbacks take the fd of the context file as an argument,
though they don't yet use it.  However, this fd will very likely be
necessary for later support of the "NARY" topology in ssh-spawner
(expected to require auxillary info in the context file).  So, the
code in ibv-conduit has been reworked as necessary to accomodate
passing it.

[1] BLCR's cr_restart utility works hard to forward most signals, but
SIGKILL (and SIGSTOP) cannot be caught.