ibv-conduit: stronger pre-checkpoint quiesce
This commit replaces an out-of-band barrier (e.g. via tcp sockets
within ssh-spawner) with an AM-based one.  Since the AM channels are
in-order, this has the side-effect of ensuring (under conditions
listed below) that all sent AMs have been received prior to the
checkpoint.

Conditions:

+ If there are multi rails in use, the in-order (FIFO) property of AMs
is lost and this change is no longer sufficient to ensure the
reception of the barrier AMs ensures the channels (now plural) are
drained.

+ If SRQ is in use, then Request and Reply are distinct channels.  So,
this change is not sufficient to ensure the Reply channel is drained.

In short this is not the final solution to the issue of draining all
AM receives prior to the checkpoint.