blcr support: two minor improvements
1) This commit hanges the "scope" of a checkpoint from "TREE" to
"PROC".  In other words this causes the checkpoint to be taken of only
the GASNet process, and NOT its children.  In particular, for the case
of a tree-based ssh-spawn (not yet supported for other reasons) this
prevents attempting to checkpoint the ssh processes.

2) Request async reporting of most checkpointing errors.  This aids
in cleaner error handling (not yet fully implemented) and already
improves detection of some problems.  For instance, the attempt to
checkpoint the ssh processes (fixed by the previous item) is now
reported as a fatal error rather than causing a hang.