ibv-conduit: fix re-initialization of SRQ
This commit removes use of a static "first" flag for controlling the
posting of receive buffers to shared receive queues.  The logic was
fine previously, but it not correct is the presence of
RE-initialization due to checkpoints.  The new logic checks if there
are any entries on the free list to determine of this is the "first"
call.