Revert "ibv-conduit: stronger pre-checkpoint q..."
This commit reverts use of gasnetc_bootstrapBarrier_ib() as a
pre-checkpoint barrier, returning to the original spawner-based
gasneti_bootstrapBarrier() call.  For reasons I don't yet understand,
the use of the AM-based barrier is leading to infrequent hangs in my
testing (which are not present with the spawner-based barrier).

The change was only a partial fix for the problem it was meant to
address.  So, reverting it just means finding the real solution (to
fully draining AM receives) is a bit more important.

This reverts commit 82454107414033bd3caba0aa7d3f0711a3c3cff3.