ibv+blcr: fix AM quiesce code
This commit ensures that Active Messages are fully quiesced when doing
a communications shutdown for checkpointing.  The basic idea is to
"grab" all the credits to ensure that every Request that was sent has
been received and the Reply (or implicit ack) has been received.  The
only complication is that we cannot stop polling until we know for
certain that we have received all incomming Requests (and have sent
the Reply or ack).  This is done using a barrier based on a special
AM Request that does NOT consume a credit, and does not elicit any
Reply or ack (which would return the credit that was not consumed).

This commit disables the credit-coallescing optimization whenever BLCR
support is enabled.  This optimization is important to the performance
of certain AM flows, and I hope to be able to restore it in a future
commit.

NOTE: only collective full-job quiesce is currently supported.
However, the same basic building blocks should be usable to construct
pair-wise quiesce operations if/when that is required.