ibv-conduit: simplify AMRDMA buffer management
This commit reimplements the management of buffers used for
AM-over-RDMA.  The management is now down using a bit "map" of
available buffers and a single "tail" value that indicates the next
bit to allocate, replacing a "head" and "tail".  (Note that allocation
*must* be done in order so the peer knows where to probe).  These two
quantities are combined into a single "tail_and_map" word to allow
atomic allocation (adding to "tail" and masking out a bit in "map").
Acknowledgements now carry a bit mask of buffers (ORed into
"tail_and_map" via atomic-add) instead of a value by whith to advance
"head".  The result is the elimination of mutex-protected logic that
had to collect acks at the target so that they would only be sent (as
a count, not a mask) when they were contiguous with the remote "head"
value.  This was the *last* mutex-protected code in any critical
path through communications code.

This change requires us to reduce the max number of buffers under
management from 32 to 16 (due to the header space available for the
"ack_bits"), but 16 has been our default depth for a long time.  That
could be raised to 19 with no changes to the header, and to as much as
24 if we impose a (quite reasonable) limit of 7 on the depth of credit
coallescing (to reduce from 8 to 3 the number of bits devoted to
credits in the header).  However, anything that is not a power-of-two
will invalidate the current use of bitwise-AND as an optimized MOD
used in manipulating the recevier's "head" and the sender's "tail".

With 4 more byes reserved for (optional) use in the header (reducing
MaxMedium), this could easily be extended up to 24 + 32 = 56 buffers
under management (noting that a 64-bit "tail_and_map" word would still
have 8 bits for the "tail" field).  That may be worth some serious
consideration if/when we are able to investigate allocation by pieces
smaller that MaxMedium (allowing, for instance, 16 MaxMediums
outstanding OR 56 small AMs in the same space).