ibv: experimental version of rdma-dissem barrier
This is an experimental version of the ibv-specific implementation
of the RDMA-based dissemination barrier.  Where the previous version
depended on the assumption that the 64-bit payload was written
atomically, this version is relaxed slightly to assume only that the
bytes are written from lowest address to highest (and written only
once).  So, atomic delivery becomes just a special case of in-order.

The specific goal of this work is to confirm that failures that had
been seen on a particular platform are due to partial (such as by host
memcpy()) delivery of the barrier payloads.