pshm: invoke cleanup on pshm barrier overflow
The logic to catch signals and fatal errors in the segment mapping
code within gasnet_init() uses an intentional overflow of the pshm
barrier counter to force termination of any processes which may be
blocked in that barrier code.  This is necessary because the full
handling of abnormal exits may not (for some conduits) be usable
until the default signal handlers are installed in gasnet_attach().

This commit ensures that the abnormal exit due to this overflow will
invoke the cleanup hook.  It is possible that this may be redundant
in many (perhaps most) circumstances, but lacking a proof that it is
*always* redundant, this change falls under "better safe than sorry".

(cherry picked from commit f194aa60a478230a192fe120e95e77afe12d00cd)