wip - separation of PSHM for segment vs AM
This commit is sufficient to allow udp conduit to use PSHM-based AMs
in a SEGMENT_EVERYTHING build, with measurable improvements in latency
and bandwidth as a result.

ALL OTHER CONDUITS ARE PROBABLY BROKEN

The per-conduit work required for mpi-conduit should be similar to
that done in this commit for udp-conduit.

For conduits with a native extended API implementation, additional
work *may* be required for selecting between the two paths (amref vs
native loopback) for the Extended API.

For smp-conduit, non-trivial additional work is anticipated to restore
(conditionally) enough of the reference code in gasnet_extended.c to
permit use of the amref implementation of the Extended API.