Fix bug 3428 using volatile sources in count0s
This commit fixes bug 3428:
  Pathscale 6.0 -O3+ misoptimizes ibv-conduit AMRDMA path

Despite the bug description, the optimization which was performed was
not technically incorrect.  In the absence of a volatile qualifier the
compiler was free to read the source data more than once.  This meant
that, in the presence of data being updated concurrently, the values
copied and the values in which zero bytes are counted could be
*different*.

This commit adds volatile qualifiers to all of the pointers which are
used to read the source data in all of the code which supports the
count0s and count0s_copy routines.  A subset of these were sufficient
to fix the reported failure mode, but the uniform use of volatile
should ensure that code is correct with even more aggressive
compilers.