testmisc: fix memcpy+count0s tests
This commit replaces
    "count0s(s,n) + memcpy(d,s,n)"
with
    "memcpy(d,s,n) + count0s(d,n)".

This better represents the manner in which the count0s should be used, since
the previous version would (for an actively changing buffer) potentially
produce a buffer and zero-count which did not match.

Additionally, this commit changes the count0s tests to add to a volatile
accumultor to prevent over-agressive optimization (discarding the entire
test!).