ssh-spawner: heterogeneous launch support
This commit implements (but does not yet document) support in
ssh-spawner for having distinct nodes run the perl script and the
"head" instance of the gasnet executable.  This allows, for instance,
ssh-based spawning where the front-end and compute nodes are not
binary compatible.  That could be a different ISA (including KNL
compute nodes w/ x86-64 front-end) or simply a system without the
necessary shared libs on the front-end.

It is worth noting that the perl script runs only on the front-end.
This is good if the compute nodes don't have perl (though of course
we do assume they have both an ssh server and client).

Example:
  gasnetrun_[conduit] -proxy='ssh node123' -n 4 ./a.out