Partial revert of commit 64ca5f6
This commit partially reverts the following:
    commit 64ca5f6e393fb4946c1203633cd01317b6226bc9

    bug 3448: x86*/Linux timers now default to posix-realtime

    A temporary workaround until bug 3446 can be addressed.

We retain the portion of the logic added by that commit which allows
for platform-specific exceptions to native-timers-by-default.
However, there are not currently any such exceptions.

Of course the ChangeLog gets a new entry reflecting the restoration of
the old default, rather then removing the old entry.