Tighten TSC calibration criterion to 2-sigma
This commit tightens the criterion used for calibration of the TSC
against the OS-provided wallclock from "1-sigma" (68% confidence) to
"1-sigma" (95% confidence).  This change makes the interpretation of
the soft and hard tolerance environment variables more consistent with
their meanings for the "cpuinfo" and user-provided calibrations.

For the current default tolerance of 0.05% this does not have any
impact on the rate of convergence (10 iters or less) on a small
sampling of systems.  However, making this change now will make any
subsequent adjustments more meaningful.