Construct tables of Client and EP objects
This commit adds statically allocated global tables of (pointers to)
the Client and EP object types; and adds _index fields to these object
types to indicate their table location.  Together these provide the
means to pass a small integer identifier "on the wire" and to convert
it back to the object type.

Macros gasneti_index2<type>() are provided to help avoid "baking in"
assumptions related to the storage.