Update the PMIx support so it properly matches PMIx v2.x.
We still have the "assert" being fired when the PMIx library calls "malloc" - this was fixed before, but that fix appears to have been lost. I pulled the PMIx code out to create a new unit test in the PMIx repository, and it passes - so I believe the PMIx integration in this branch is correct.

Just need a little help to resolve the "assert". Note that this was tested so far with only the OFI transport, using the sockets provider.

Signed-off-by: Ralph Castain <rhc@open-mpi.org>