Initial support for Cray compilers on AARCH64
This commit makes necessary (but not sufficient) changes to allow
compilation of GASNet using Cray's compilers for AARCH64 (aka ARMv8).
The changes extend default {C,CXX}FLAGS and a bug work-around for the
x86-64 compiler to the AARCH64 version.

Notably missing are memory fences and atomics.
There is also no suitable cross-configure script at this time.