ofi: Make async long AMs truly asynchronous
In this commit, asynchronous long AMs and their associated RMA data
transfers will now carry an ID that is an index into a hash
table on the target, used to reconstruct the message. This allows the
sender to return immediately instead of stalling for the data transfer.

The ID is 32 bits, with the top 16 bits representing the node
and the bottom 16 bits representing a sequence number. This scheme
limits the max number of nodes to 2^16