atomics: no native atomics for unknown compilers
This commit resolves
   Bug 3668 - RFE: slow-atomics for "mismatched" compilers

This commit ensures that when using an "unknown" compiler (not CC, CXX
or MPI_CC probed at configure time) calls to atomics will only resolve
to use of the "generic" (mutex based) or "slow" (function call based)
implementations.  Selection is made independently for 32- and 64-bit
as needed to allow for ILP32 platforms using native for the former and
generic for the later.

This is the safest approach to atomics on unknown compilers, but there
are at some ways this might be improved in the future:

  1. For "compatible" compilers (such as later version of same family)
     we may wish to allow compilation of the native atomics.
  2. For the case of "special" atomics (only Sun C currently), the
     slow path is currently two function calls deep, with the leaf
     functions providing just a subset of the full operations and the
     other functions implementing the full set in terms of the leaves.
  3. As with "special" atomics, use of "os" atomics may result in a
     second level of function call (but "is" is not the current
     default for any platform).