timers/membars: support xlc as "unknown compiler"
This commit adds (or modifies) preprocessor logic in the timers and
membars support to allow use of an IBM XL Compiler which was not
probed at configure time.

This is *not* a fully general solution for use of "SLOW" timers or
membars for unknown compliers.  This is a solution only for IBM XL
Compilers which are particularly problematic because support for
GNU-style and XLC-style inline assembly vary by version.