atomics: correct the handling of "special" atomics
This commit distinguishes the "SPECIAL" atomics as a distinct class of
atomics implementation (rather than considering them as NATIVE) for the
purposes of determining compatibility for unknown compilers and
mixed-family builds.  However, the configure output and config string
continue to consider them to be "native" implementations.

In addition to adding logic for distinguishing SPECIAL atomics, this
commit expands their documentation and the documentation on
compatibility of atomics implementations.