coll-ex: raise limit on Reduce "payload"
This commit provides a new environment variable to set a minimum
"per-log_2(ranks)" size for the eager scratch space, and sets its
default value to 32 bytes.  Note that due to the presence of a
pre-existing linear scaling factor, this only results in increased
storage for jobs of size exactly 3 or 5 ranks.

This ensures that gex_Coll_ReduceNB() will support `dt_sz*dt_cnt` no
less than 32 bytes in all configurations (up from 24).  More
importantly, it provides the means to raise the value when needed.

The new env var (like its peers to set the min and linear scaling
factor) is UNDOCUMENTED.  The mechanism for sizing of eager buffer
space is likely to change with GEX collectives rewrites, making
documentation at this time unwise.

Since the env var is not documented in the normal places, the error
message for too-large `dt_sz*dt_cnt` is updated to mention it as a
means "IN THIS RELEASE ONLY" to raise the limit.

The specification is updated to describe the current limitation.

Logic in testcollperf meant to mirror the computation of the limit has
been updated as well.