coll-ex: fix alg selection computation
This commit corrects an erroneous computation in the selection of an
algorithm for gex_Coll_ReduceToOneNB().  Since there is only *one*
implementation, this means that for arguments too large to support one
would get an assertion failure (DEBUG) or buffer overruns (NDEBUG)
instead of the more desirable explanatory fatalerror message.