coll: remove gasnet_coll_init()
This commit (finally) removes the legacy gasnet_coll_init() function
and the remaining calls from the tests.

It had already ceased to perform a role in registering "images", and
with the elimination of the last calls to gasnet_coll_reduce(), it is
no longer needed to support registration of reduction functions.