coll: add max_radix to local-view geom
This commit adds the maximum radix (a.k.a. out-degree) of a tree to
the "local view" geometry struct.  Though this is actually a global
property of the tree, it is one that is required to select a reduction
algorithm, because it determines the largest volume of data to be
received at any tree node.