ibv: address bug 3693
This commit implements an optional hook in gasneti_SegmentAttach() to
permit a conduit to (for instance) attempt pinning before the segment
is "committed", and to request a smaller segment size if the attempt
were to fail.  This hook is implemented for ibv-conduit to deal with
memory registration failures such as reported in bug 3693.