gemini/aries: add "notify" field to AM wire format
This commit adds an 8-bytes field to the on-the-wire format used for
AM communications, which will carry the value currently known as the
"notify word".  This will allow for protocol changes which will need
to operate without use of the "notify ring".

This commit only adds the field and adjusts the Max Medium (LUB
actually) configure logic to deal with the enlarged header.  This
commit does not yet read or write the new field.

This change lowers gex_AM_LUB{Reqeust,Reply}Medium() by 8-bytes.
However, it is hoped that this will be more than offset by a later
increase (ideally to a value as large as 64K minus header).