gemini/aries: move notify word to packet header
This commit relocates the primary storage of the "notify word" by an
AM initiator from the `sync_flag_value` field of the post descriptor
to the newly-added `header` field in the AM packet itself, with the
result that these 8 bytes are now sent redundantly (once in to the
remote notify ring, as always, and now in the payload as well).
This will permit additional changes to AM protocols in which the
notify ring is not used (and thus only a single copy is sent).