gemini/aries: use GNI_EpSetEventData to set cqdata
This commit introduces use of GNI_EpSetEventData() in the AM injection
paths to manipulate the value transmitted as the "Remote Instance ID"
and thus received as the CQ Data at the target.

This initial commit is a "no op" in that it pointlessly sets the value
to its default.  However, future commits will use this to encode data
used by the AM protocols.