gemini/aries: move AMReply traffic off notify
This commit revises the protocol used for all AM Reply traffic
(independent of Eager vs Rendezvous) to use a simple FMA_PUT (rather
than FMA_PUT_W_SYNCFLAG).

The 32 bits of "instance id" delivered to the remote CQ are used
(instead of the "notify word") to pass the initiator_slot, from which
the recipient can recover everything required to process the Reply.

This halves the size of the notify ring, and removing the dependence
on the "W_SYNCFLAG" opens the possibility of RDMA (rather than FMA)
for Replies in the future (to accommodate larger payloads).