gemini/aries: convert AM rendezvous to CQWRITE
This commit converts the protocol used for an AM Rendezvous (a
Request to be fetch by the target using a Get) to use a CQWRITE rather
than an FMA_PUT to the notify ring.  The 48-bits of data delivered to
the remote CQ entry is sufficient to initiate the Get of the full
message.

This commit also factors logic used for posting a CQWRITE.

In theory, this commit is sufficient allow the notify ring to be
eliminated when the Rendezvous protocol is active.  However, that
modification has been omitted because it has become clear that
upcoming changes will entirely eliminate the notify ring (for the
Eager protocol as well).