gemini/aries: move AMRequest traffic off notify
This commit revises the protocol used for Eager AM Request traffic
to use a simple FMA_PUT rather than FMA_PUT_W_SYNCFLAG.

The 32 bits of "instance id" delivered to the remote CQ are used
(instead of the "notify word") to pass the source_id and target_slot
form which the recipient can recover the address at which the
Request has been delivered.

Removing the dependence on the "W_SYNCFLAG" opens the possibility of
RDMA (rather than FMA) for Requests in the future (to accommodate
larger payloads).

The "notify ring" is now unused and will be removed in a future
commit.