gemini/aries: align notify and cqdata constants
This commit swaps the order of two values in one enum, and removes an
unused value from another.  This makes the values of textually similar
enum values equal (with a small possibility of some code gen
improvements where they are used together).

There are no semantic changes associated w/ the renumbering of enums.