Remove support for IBM XL compilers for big-endian
Historically we've supported IBM's XL compilers on big-endian PPC
(with their unique inline assembly syntax) on AIX, macOS and Linux.
Our support for AIX was abandoned some time ago, and Apple stopped
supporting PPC even before that.  That left only Linux on big-endian
PPC64.  With our retirement of pami-conduit and BG/Q support we no
longer have any target platform with this compiler, and we expect to
lose access at the end of 2019.

Note that XL compilers for little-endian use a different (clang-based)
front-end and are still supported.