Add MCS locks to gasnet_syncops.h
This commit adds an implementation of MCS locks to `gasnet_syncops.h`.

Except as noted below, this implementation is "fair" (fifo queueing)
and features a `trylock` operation which "fails" in the presence of
both a lock-holder and lock waiters.  This is a key difference
relative to `pthread_mutex_trylock()`, which can "out race" a
concurrent call to `pthread_mutex_lock()`.

On platforms using generic atomics for pointer-width integers, we fall
back to use of `gasneti_mutex_t` (based on pthread mutexes).  This
means the fairness and lowered-priority trylock are not available on
such platforms.  Currently the only cases are "marginal" ones.

Additionally, in a `!GASNETI_USE_TRUE_MUTEXES` build (where no
exclusion is required) `gasneti_mutex_t` is used for its error
checking properties.