MIGHT OR MIGHT NOT MERGE - STILL TBD Things look good on Cori, but previous measurements on ARM64/smp suggest that this change would be bad on ARM64-based Cray XC. However, we lack the means to test.
--
aries: use MCS lock to protect AM request maps

When requesting AM injection resources in aries-conduit, there is a
need to serialize threads which are sending to the same peer.  Without
serialization, a series of small buffer requests could starve a large
buffer request.  This commit uses a `gasneti_mcslock_t` to replace a
hand-rolled construct for this serialization.

Previously, the AM buffer lock (protecting shared / peer-independent
state) was also used to protect a boolean `remote_request_lock` field
in per-peer state.  If that was found to be "held" then a thread would
release the AM buffer lock to spin until it was "free", and then
contend on the AM buffer lock once again.  This had the undesirable
side-effect of adding additional traffic to this AM buffer lock.
However, prior benchmarking had shown that adding a per-peer mutex was
slower under contention than this construct.

Benchmarking shows that use of a per-peer `gasneti_mcslock_t` for the
per-peer serialization performs as well as the prior mechanism in the
case of multiple threads sending to the same peer.  Since the AM
buffer lock is no longer used to provide this serialization, it should
cycle less frequently and have a shorter critical section.
Additionally, with per-peer serialization now preceding acquisition of
the AM buffer lock, the number of threads that may contend for it at
any given time may be reduced.

TL;DR:  We had previously chosen not to use a combination of "local"
and "global" locks where it would have made sense to do so, because an
ugly hand-rolled construct using only the global lock was faster.
With use of an MCS lock as the "local" one, we now make the sensible
distinction between the local and global state with no observed loss
of performance.