ibv: implement NPAM Medium
This commit implements Negotiate-Payload AM Medium (both Request and
Reply).  For the gasnet-provided buffer case this implementation is
zero-copy and does not use malloc/free.

Relative to FPAM, this preserves the inline-send optimization (using
space in the thread-specific "sd" rather than on the stack).  However,
the use of gather-on-send for in-segment client-provided payload
buffer has not been implemented, pending at least some discussion of
its benefits.