Improved diagnostic for NULL/INVALID object types
This commit updates all of the "object import" functions to check for
the corresponding `GEX_[TYPE]_INVALID` (if any) or `NULL`.  This
provides an assertion to replace the SEGV which would otherwise
occur when callers use an unchecked return value.  It is worth noting
that the import functions only exist in DEBUG builds.

Callers will need to check for INVALID *prior* to import if/when an
"INVALID" object handle is a valid input, or one desires a diagnostic
that is either more descriptive or is active in NDEBUG builds.