Implement --(en|dis)able-assert
This commit begins separation of assettions from opt/dbg compiler flags
by adding `--(en|dis)able-assert` to `configure`.  As with trace and
stats support, the default is to follow `--enable-debug`.

The case of `--enable-assert` without `--enable-debug` is currently
broken, due to multiple cases of assertions which consume variables (and
call helper functions?) conditionally defined only in `GASNET_DEBUG`
builds.  Additionally, it is suspcted that assertions may mis-fire due
to state that is only maintained accurately in a `GASNET_DEBUG` build.