Implement gex_Segment_EP_Bind()
This commit implements `gex_Segment_EP_Bind()` as documented in the
2020.6.1 API proposal.

There is a single "reference" implementation currently used by all
conduits.  However, a mechanism for conduit-specific override is
provided, including the ability to call the reference implementation.

Additionally, the implementation of `gasneti_segmentAttach()` is
slightly modified (operations reordered and comments added) to better
identify how its operations correspond to the user-facing APIs.  At
some point this function might actually be reimplemented in terms of
the low-level functions implementing the user-facing APIs.