Implement and document gex_Segment_Publish()
This commit implements and documents `gex_Segment_Publish()`.

There is a reference implementation currently used by conduits using
the reference extended API code, and a mechanism by which conduit with
a native extended API implementation provide their overrides (calling
the reference implementation for much/most of the work).

It is planned that in the future the conduit-specific logic will be
more tightly integrated with the conduit-independent logic, in
particular to avoid redundant communication of `gex_EP_Location_t`
data.  However, sharing of common code with the aux segment
registration and gex_Segment_Attach in several conduits makes the
changes required non-trivial.  Additionally, both ibv and ucx have
variable-length registration key data which renders a naive
"EXTRA_FIELDS" approach insufficient on it own.