ucx: revise gasnetc_segment_exchange()
This commit revises `gasnetc_segment_exchange()` to work on teams
other than the primordial one, while retaining its (so far unique)
ability to operate on the aux segment.

This does not yet generalize this function for multi-ep support as has
been done in the same-named functions in other conduits.  However,
until Create_EP is implemented, the current support for 0 or 1
endpoints is sufficient.