pshm: move xpmem segment ids into shared memory
This commit uses the recently added hooks for pshm "private data"
to replace the replicated table of xpmem segment ids with a single
table in shared space.

A intra-node exchange operation in the "publish" step of PSHM wire-up
is replaced with writes and an intra-node barrier.  However, this
barrier is made optional to enable later work which will include
non-collective cross-mapping of segments.

It is worth noting that xpmem is the only PSHM implementation which
has a non-empty "publish" step.  Other implementations use a shared
name from which all the segment identifiers can be derived without
communication.