Implement gex_TM_Pair()
The commit implements `gex_TM_Pair()` as documented in the 2020.6.1
API proposal.

Testing with the new `testtmpair` has verified that the implementation
is sufficient to accept a "pair" naming the primordial endpoints in
the expected APIs.  This includes all conduits (smp, udp, mpi, ibv,
ucx, aries, ofi), in both FAST and EVERYTHING modes where applicable.