Maintain a per-client table of endpoints
This commit adds a table of endpoints to the client object, with
modifications to the client and endpoint allocation functions to
maintain that table.

Additionally, this commit lowers the default `GASNET_MAXEPS` from 2^12
to just 1, to accurately reflect the current implementation
limitation.  This limit is expected to be raised independently by
each conduit as multi-endpoint support is added.

Connected to the lowered `GASNET_MAXEPS`, an assertion upon exceeding
the limit has been replaced with a more informative fatalerror.