Implement gasneti_import_tm_nonpair()
This commit provides an alternative to `gasneti_import_tm()` with the
added property of checking (in a DEBUG build) that the argument is NOT
a TM Pair, and printing an explanatory error message otherwise.

Additionally, a good-faith effort has been put into using the new
variant in all places where a Pair is invalid and the code does not
already handle or exclude the possibility.  However, it is possible
some have been missed.